+++ MojoUnity-Shake/ChangeLog.md
@@ -1,3 +1,10 @@
+## v1.1
+_`2021.8.9 UTC+8 15:30`_
+
+* Remove Smooth type.
+* Optimize `BurstCompile` with `math` API.
+* Fix shake range mapping from [0, 1] to [-1, 1].
+
 ## v1.0
 _`2021.4.23 UTC+8 11:15`_